+++ package.json
@@ -63,8 +63,6 @@
   "dependencies": {
     "@elastic/elasticsearch": "7.9.0",
     "@elastic/elasticsearch-mock": "0.3.0",
-    "@material-ui/core": "4.11.0",
-    "@material-ui/lab": "4.0.0-alpha.56",
     "@wordpress/wordcount": "2.11.0",
     "body-parser": "1.19.0",
     "bull": "3.18.0",
